+++ nflow-engine/src/main/java/io/nflow/engine/internal/executor/DispatcherExceptionAnalyzer.java
@@ -1,0 +1,5 @@
+package io.nflow.engine.internal.executor;
+
+public interface DispatcherExceptionAnalyzer {
+  DispatcherExceptionHandling analyze(Exception e);
+}